WIP: adding crust as model

+++ frospy/core/splittingfunc/read.py
@@ -76,7 +76,7 @@ def read_cst(setup=None, modes=None, cfile=None, modes_dir=None, R=-0.2,
     """
 
     if setup is not None or modes_dir is not None:
-        if cfile in ('S20RTS', 'S40RTS', 'SP12RTS', 'QRFSI12'):
+        if cfile in ('S20RTS', 'S40RTS', 'SP12RTS', 'QRFSI12', 'CRUST'):
             if len(setup.modes_cc) > 0 and len(setup.modes_sc) == 0:
                 allmodes = read_modes()
                 for _m in setup.modes_cc.keys():
@@ -125,7 +125,7 @@ def read_cst(setup=None, modes=None, cfile=None, modes_dir=None, R=-0.2,
     elif cfile == 'RR':
         cst, dst, cst_errors, dst_errors = read_cst_RR(modesin, modes_ccin,
                                                        verbose=verbose)
-    elif cfile in ('S20RTS', 'S40RTS', 'SP12RTS', 'QRFSI12'):
+    elif cfile in ('S20RTS', 'S40RTS', 'SP12RTS', 'QRFSI12', 'CRUST'):
         cst, dst = read_cst_S20RTS(modesin=modesin, modes_ccin=modes_ccin,
                                    setup=setup, modes_dst=modes_scin_dst,
                                    R=R, model=cfile,
@@ -1411,6 +1411,14 @@ def read_cst_S20RTS(modesin, modes_ccin, setup=None, bin_path=None,
         _maxcdeg = 12 # crust model
         _maxddeg = 12 # dst model
 
+    if model == 'CRUST':
+        cstS20RTS = None
+        cc_cstS20RTS = None
+        dstS20RTS = None
+        cc_dstS20RTS = None
+        _maxmdeg = 12 # cst model
+        _maxcdeg = 12 # crust model
+        _maxddeg = 12 # dst model
     sc_modes, cc_modes = get_mode_names(modesin, modes_ccin)
     sc_cdeg, sc_ddeg, cc_cdeg, cc_ddeg = get_mode_deg(modesin, modes_ccin)
 
@@ -1446,7 +1454,7 @@ def read_cst_S20RTS(modesin, modes_ccin, setup=None, bin_path=None,
             os.system('echo "%03d %s %03d" >> modes.in' % (int(m[0]),
                                                            m[1].lower(),
                                                            int(m[2])))
-            if model != 'SP12RTS':
+            if model != 'SP12RTS' and model != 'CRUST':
                 # S20RTS prediction
                 for s in np.arange(0, int(s_max)+1, 2):
                     # only input coupling degrees, No degree higher then 20
@@ -1591,22 +1599,23 @@ def read_cst_S20RTS(modesin, modes_ccin, setup=None, bin_path=None,
                         ccdegs = max_cc_degrees(max_cc_input)
 
                         # S20RTS cc prediction
-                        for s in np.arange(sdeg[0], int(sdeg[1])+1, 2):
-                            # only input coupling degrees
-                            if s not in ccdegs:
-                                continue
-                            os.system('echo "%s" > input' % s)
-                            res = subprocess.Popen('%s < input' % cc_cstS20RTS,
-                                                   shell=True,
-                                                   stdout=subprocess.PIPE,
-                                                   stderr=subprocess.PIPE)
+                        if model != 'CRUST':
+                            for s in np.arange(sdeg[0], int(sdeg[1])+1, 2):
+                                # only input coupling degrees
+                                if s not in ccdegs:
+                                    continue
+                                os.system('echo "%s" > input' % s)
+                                res = subprocess.Popen('%s < input' % cc_cstS20RTS,
+                                                       shell=True,
+                                                       stdout=subprocess.PIPE,
+                                                       stderr=subprocess.PIPE)
 
-                            output, error = res.communicate()
-                            os.system('cat mcst.dat >> mcst-S20RTS.dat')
-                            os.remove('mcst.dat')
-                        with open('mcst-S20RTS.dat', 'r') as fh:
-                            c_s20rts_tmp = np.genfromtxt(fh)
-                        os.remove('mcst-S20RTS.dat')
+                                output, error = res.communicate()
+                                os.system('cat mcst.dat >> mcst-S20RTS.dat')
+                                os.remove('mcst.dat')
+                            with open('mcst-S20RTS.dat', 'r') as fh:
+                                c_s20rts_tmp = np.genfromtxt(fh)
+                            os.remove('mcst-S20RTS.dat')
                         # CRUST cc prediction
                         if include_CRUST is True:
                             for s in np.arange(sdeg[0], int(sdeg[1])+1, 2):
@@ -1634,6 +1643,8 @@ def read_cst_S20RTS(modesin, modes_ccin, setup=None, bin_path=None,
                         # os.remove('mcst.dat')
             else:
                 cc_modesin = None
+
+    # dst calculation here?
     if modes_dst is None:
         dst = None
         sc_coeff_dst = None
